# specs/001-leave-management-system/contracts/api.yaml
openapi: 3.0.3
info:
  title: Leave Management System API
  description: API for managing leave requests, approvals, documents, and notifications
  version: 1.0.0
servers:
  - url: https://your-app.vercel.app/api
    description: Production
paths:
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                department:
                  type: string
      responses:
        "201":
          description: User created
        "400":
          description: Invalid input
...
  /leaves/{id}/approve:
    post:
      summary: Approve leave request
      tags: [Leaves]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comments:
                  type: string
      responses:
        "200":
          description: Leave approved
        "403":
          description: Not authorized
  /leaves/{id}/reject:
    post:
      summary: Reject leave request
      tags: [Leaves]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comments]
              properties:
                comments:
                  type: string
      responses:
        "200":
          description: Leave rejected
        "400":
          description: Comments required
...
  /documents:
    get:
      summary: Get documents
      tags: [Documents]
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      summary: Upload document
      tags: [Documents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title, category]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                expiryDate:
                  type: string
                  format: date
                accessLevel:
                  type: string
                  enum: [PUBLIC, ROLE_RESTRICTED, PRIVATE]
      responses:
        "201":
          description: Document uploaded
        "400":
          description: Invalid file
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
    UserRole:
      type: string
      enum: [EMPLOYEE, MANAGER, HR, ADMIN]
...
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        fileName:
          type: string
        fileSize:
          type: number
        mimeType:
          type: string
        accessLevel:
          type: string
          enum: [PUBLIC, ROLE_RESTRICTED, PRIVATE]
        uploadedBy:
          $ref: "#/components/schemas/User"
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT